# contracts/local-agent.openapi.json
{
  "components": {
    "schemas": {
...
      "ClientGrantRenewalInput": {
        "additionalProperties": false,
        "properties": {
          "grant_id": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Grant Id",
            "type": "string"
          },
          "pairing_id": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Pairing Id",
            "type": "string"
          }
        },
        "required": [
          "pairing_id",
          "grant_id"
        ],
        "title": "ClientGrantRenewalInput",
        "type": "object"
      },
      "ClientGrantRenewalResponse": {
        "additionalProperties": false,
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "grant": {
            "$ref": "#/components/schemas/ClientGrantResponse"
          },
          "token_type": {
            "const": "DPoP",
            "default": "DPoP",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "grant",
          "access_token",
          "expires_at"
        ],
        "title": "ClientGrantRenewalResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "correlation_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._~-]*$",
            "title": "Correlation Id",
            "type": "string"
          },
          "detail": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Detail",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProblemDetailsResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_code": {
            "$ref": "#/components/schemas/ProblemCode"
          },
          "instance": {
            "pattern": "^urn:inari:request:[A-Za-z0-9][A-Za-z0-9._~-]{0,127}$",
            "title": "Instance",
            "type": "string"
          },
          "message_key": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_]*$",
            "title": "Message Key",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "status": {
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "type": {
            "pattern": "^urn:inari:problem:v1:[a-z][a-z0-9_]*$",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "error_code",
          "message_key",
          "retryable",
          "correlation_id"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Inari",
    "version": "1.20.0a1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/pairing/v1/client-grants/renew": {
      "post": {
        "operationId": "renew_client_grant",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientGrantRenewalInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientGrantRenewalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details",
            "headers": {
              "Cache-Control": {
                "description": "Prevents storage of the nonce challenge.",
                "schema": {
                  "enum": [
                    "no-store"
                  ],
                  "type": "string"
                }
              },
              "DPoP-Nonce": {
                "description": "Fresh single-use nonce for one DPoP retry.",
                "schema": {
                  "type": "string"
                }
              },
              "WWW-Authenticate": {
                "description": "DPoP challenge with the `use_dpop_nonce` error.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "410": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          }
        },
        "summary": "Renew Client Grant",
        "tags": [
          "pairing"
        ]
      }
    },
    "/pairing/v1/requests": {
      "post": {
        "operationId": "create_client_pairing_request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairingRequestCreateInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairingRequestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details",
            "headers": {
              "Cache-Control": {
                "description": "Prevents storage of the nonce challenge.",
                "schema": {
                  "enum": [
                    "no-store"
                  ],
                  "type": "string"
                }
              },
              "DPoP-Nonce": {
                "description": "Fresh single-use nonce for one DPoP retry.",
                "schema": {
                  "type": "string"
                }
              },
              "WWW-Authenticate": {
                "description": "DPoP challenge with the `use_dpop_nonce` error.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "410": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          }
        },
        "summary": "Create Client Pairing Request",
        "tags": [
          "pairing"
        ]
      }
    },
    "/pairing/v1/requests/{request_id}": {
      "get": {
        "operationId": "get_client_pairing_request",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairingRequestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details",
            "headers": {
              "Cache-Control": {
                "description": "Prevents storage of the nonce challenge.",
                "schema": {
                  "enum": [
                    "no-store"
                  ],
                  "type": "string"
                }
              },
              "DPoP-Nonce": {
                "description": "Fresh single-use nonce for one DPoP retry.",
                "schema": {
                  "type": "string"
                }
              },
              "WWW-Authenticate": {
                "description": "DPoP challenge with the `use_dpop_nonce` error.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "410": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          }
        },
        "summary": "Get Client Pairing Request",
        "tags": [
          "pairing"
        ]
      }
    },
    "/pairing/v1/requests/{request_id}/admit": {
      "post": {
        "operationId": "admit_client_pairing",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairingAdmissionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairingAdmissionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details",
            "headers": {
              "Cache-Control": {
                "description": "Prevents storage of the nonce challenge.",
                "schema": {
                  "enum": [
                    "no-store"
                  ],
                  "type": "string"
                }
              },
              "DPoP-Nonce": {
                "description": "Fresh single-use nonce for one DPoP retry.",
                "schema": {
                  "type": "string"
                }
              },
              "WWW-Authenticate": {
                "description": "DPoP challenge with the `use_dpop_nonce` error.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "410": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          }
        },
        "summary": "Admit Client Pairing",
        "tags": [
          "pairing"
        ]
      }
    },
    "/pairing/v1/requests/{request_id}/cancel": {
      "post": {
        "operationId": "cancel_client_pairing_request",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairingRequestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details",
            "headers": {
              "Cache-Control": {
                "description": "Prevents storage of the nonce challenge.",
                "schema": {
                  "enum": [
                    "no-store"
                  ],
                  "type": "string"
                }
              },
              "DPoP-Nonce": {
                "description": "Fresh single-use nonce for one DPoP retry.",
                "schema": {
                  "type": "string"
                }
              },
              "WWW-Authenticate": {
                "description": "DPoP challenge with the `use_dpop_nonce` error.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "410": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          }
        },
        "summary": "Cancel Client Pairing Request",
        "tags": [
          "pairing"
        ]
      }
    },
    "/pairing/v1/requests/{request_id}/decision": {
      "post": {
        "operationId": "decide_client_pairing_request",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairingDecisionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairingRequestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details",
            "headers": {
              "Cache-Control": {
                "description": "Prevents storage of the nonce challenge.",
                "schema": {
                  "enum": [
                    "no-store"
                  ],
                  "type": "string"
                }
              },
              "DPoP-Nonce": {
                "description": "Fresh single-use nonce for one DPoP retry.",
                "schema": {
                  "type": "string"
                }
              },
              "WWW-Authenticate": {
                "description": "DPoP challenge with the `use_dpop_nonce` error.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "410": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RFC 9457 Problem Details"
          }
        },
        "summary": "Decide Client Pairing Request",
        "tags": [
          "pairing"
        ]
      }
    },
...
  }
}